adapt test cases
