+++ misc/services/camel/de-metas-camel-edi/src/test/resources/de/metas/edi/esb/desadvexport/ecosio/DESADV_1023358_1611220554035_expected_output.xml
@@ -15,47 +15,47 @@
     <EDI_ExportStatus>D</EDI_ExportStatus>
     <MovementDate>2021-01-20+01:00</MovementDate>
     <POReference>6666666666</POReference>
-    <EDI_Exp_Desadv_Pack>
-        <GTIN_LU_PackingMaterial>EURO</GTIN_LU_PackingMaterial>
-        <IPA_SSCC18>123456789012345678</IPA_SSCC18>
-        <M_HU_PackagingCode_LU_Text>ISO</M_HU_PackagingCode_LU_Text>
-        <EDI_Exp_Desadv_Pack_Item>
-            <BestBeforeDate>2021-03-10+01:00</BestBeforeDate>
+    <EDI_Exp_DesadvLine>
+        <C_UOM_ID>
+            <Name>Kollo</Name>
+            <UOMSymbol>Kollo</UOMSymbol>
+            <X12DE355>COLI</X12DE355>
+        </C_UOM_ID>
+        <EDI_Exp_DesadvLine_Pack>
+            <IPA_SSCC18>123456789012345678</IPA_SSCC18>
             <LotNumber>19</LotNumber>
             <QtyCU>1</QtyCU>
             <QtyCUsPerLU>1</QtyCUsPerLU>
             <QtyTU>1</QtyTU>
-            <EDI_DesadvLine_ID>
-                <C_UOM_ID>
-                    <Name>Kollo</Name>
-                    <UOMSymbol>Kollo</UOMSymbol>
-                    <X12DE355>COLI</X12DE355>
-                </C_UOM_ID>
-                <EDI_DesadvLine_ID>1000000</EDI_DesadvLine_ID>
-                <Line>100</Line>
-                <M_Product_ID>
-                    <Name>ProductName</Name>
-                    <Value>123457</Value>
-                    <Volume>0</Volume>
-                    <Weight>5.00</Weight>
-                </M_Product_ID>
-                <ProductDescription>ProductDescription</ProductDescription>
-                <ProductNo>ProductNo</ProductNo>
-                <QtyEntered>1.00</QtyEntered>
-                <IsSubsequentDeliveryPlanned>N</IsSubsequentDeliveryPlanned>
-                <GTIN>1234567890125</GTIN>
-                <PriceActual>19.900</PriceActual>
-                <EanCom_Invoice_UOM>KGM</EanCom_Invoice_UOM>
-                <InvoicableQtyBasedOn>CatchWeight</InvoicableQtyBasedOn>
-                <OrderPOReference>6666666666</OrderPOReference>
-                <QtyDeliveredInInvoiceUOM>5.16</QtyDeliveredInInvoiceUOM>
-                <QtyDeliveredInUOM>1</QtyDeliveredInUOM>
-                <OrderLine>100</OrderLine>
-            </EDI_DesadvLine_ID>
-            <GTIN_TU_PackingMaterial>BPARTNER_GTIN</GTIN_TU_PackingMaterial>
+            <BestBeforeDate>2021-03-10+01:00</BestBeforeDate>
+            <M_HU_PackagingCode_LU_Text>ISO1</M_HU_PackagingCode_LU_Text>
             <M_HU_PackagingCode_TU_Text>COLI</M_HU_PackagingCode_TU_Text>
-        </EDI_Exp_Desadv_Pack_Item>
-    </EDI_Exp_Desadv_Pack>
+            <C_UOM_ID>
+                <Name>Kollo</Name>
+                <UOMSymbol>Kollo</UOMSymbol>
+                <X12DE355>COLI</X12DE355>
+            </C_UOM_ID>
+        </EDI_Exp_DesadvLine_Pack>
+        <Line>100</Line>
+        <M_Product_ID>
+            <Name>ProductName</Name>
+            <Value>123457</Value>
+            <Volume>0</Volume>
+            <Weight>5.00</Weight>
+        </M_Product_ID>
+        <ProductDescription>ProductDescription</ProductDescription>
+        <ProductNo>ProductNo</ProductNo>
+        <QtyEntered>1.00</QtyEntered>
+        <IsSubsequentDeliveryPlanned>N</IsSubsequentDeliveryPlanned>
+        <GTIN>1234567890125</GTIN>
+        <PriceActual>19.900</PriceActual>
+        <EanCom_Invoice_UOM>KGM</EanCom_Invoice_UOM>
+        <InvoicableQtyBasedOn>CatchWeight</InvoicableQtyBasedOn>
+        <OrderPOReference>6666666666</OrderPOReference>
+        <QtyDeliveredInInvoiceUOM>5.16</QtyDeliveredInInvoiceUOM>
+        <QtyDeliveredInUOM>1</QtyDeliveredInUOM>
+        <OrderLine>100</OrderLine>
+    </EDI_Exp_DesadvLine>
     <Bill_Location_ID>
         <GLN>1234567890126</GLN>
         <Name>InvoiceTo-Name EDI-GLN</Name>
