--- misc/services/camel/de-metas-camel-edi/src/test/resources/de/metas/edi/esb/desadvexport/ecosio/DESADV_noPacks_expected_output.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<EDI_Exp_Desadv AD_Client_Value="InterCheese AG" ReplicationMode="1" ReplicationType="M" Version="*" SequenceNo="1017175" TrxName="">
-    <C_BPartner_ID>
-        <EdiRecipientGLN>1234567890123</EdiRecipientGLN>
-        <Name>BuyerName</Name>
-        <Value>12345</Value>
-    </C_BPartner_ID>
-    <C_BPartner_Location_ID>
-        <GLN>1234567890124</GLN>
-        <Name>BuyerName EDI-GLN</Name>
-    </C_BPartner_Location_ID>
-    <DateOrdered>2021-01-19+01:00</DateOrdered>
-    <DocumentNo>9999999</DocumentNo>
-    <EDI_Desadv_ID>1016956</EDI_Desadv_ID>
-    <EDI_ExportStatus>D</EDI_ExportStatus>
-    <MovementDate>2021-01-20+01:00</MovementDate>
-    <POReference>6666666666</POReference>
-    <Bill_Location_ID>
-        <GLN>1234567890126</GLN>
-        <Name>InvoiceTo-Name EDI-GLN</Name>
-    </Bill_Location_ID>
-    <C_Currency_ID>
-        <CurSymbol>CHF</CurSymbol>
-        <ISO_4217Numeric>756</ISO_4217Numeric>
-        <ISO_Code>CHF</ISO_Code>
-    </C_Currency_ID>
-    <HandOver_Location_ID>
-        <GLN>1234567890127</GLN>
-        <Name>HandOver-Name</Name>
-    </HandOver_Location_ID>
-    <DropShip_Location_ID>
-        <GLN>1234567890128</GLN>
-        <Name>DropShip-Name</Name>
-        <Setup_Place_No>1234</Setup_Place_No>
-    </DropShip_Location_ID>
-    <InvoicableQtyBasedOn>CatchWeight</InvoicableQtyBasedOn>
-    <EDI_Exp_DesadvLineWithNoPack>
-        <EDI_DesadvLine_ID>
-            <C_UOM_ID>
-                <Name>Stück</Name>
-                <UOMSymbol>Stk</UOMSymbol>
-                <X12DE355>PCE</X12DE355>
-            </C_UOM_ID>
-            <EDI_DesadvLine_ID>1000000</EDI_DesadvLine_ID>
-            <Line>10</Line>
-            <M_Product_ID>
-                <Name>Convenience Salat 250g</Name>
-                <Value>P002737</Value>
-                <Volume>0</Volume>
-                <Weight>0.250000000000</Weight>
-            </M_Product_ID>
-            <ProductDescription>Convenience Salat 250g</ProductDescription>
-            <QtyEntered>20</QtyEntered>
-            <IsSubsequentDeliveryPlanned>N</IsSubsequentDeliveryPlanned>
-            <PriceActual>2.00</PriceActual>
-            <EanCom_Invoice_UOM>PCE</EanCom_Invoice_UOM>
-            <InvoicableQtyBasedOn>Nominal</InvoicableQtyBasedOn>
-            <OrderPOReference>poRef</OrderPOReference>
-            <QtyItemCapacity>10</QtyItemCapacity>
-            <QtyDeliveredInInvoiceUOM>0</QtyDeliveredInInvoiceUOM>
-            <QtyDeliveredInUOM>0</QtyDeliveredInUOM>
-            <OrderLine>10</OrderLine>
-        </EDI_DesadvLine_ID>
-    </EDI_Exp_DesadvLineWithNoPack>
-</EDI_Exp_Desadv>
