--- misc/services/camel/de-metas-camel-edi/src/test/resources/de/metas/edi/esb/desadvexport/ecosio/DESADV_with_and_without_packs.xml
@@ -1,186 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ #%L
-  ~ de-metas-camel-edi
-  ~ %%
-  ~ Copyright (C) 2022 metas GmbH
-  ~ %%
-  ~ This program is free software: you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as
-  ~ published by the Free Software Foundation, either version 2 of the
-  ~ License, or (at your option) any later version.
-  ~
-  ~ This program is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public
-  ~ License along with this program. If not, see
-  ~ <http://www.gnu.org/licenses/gpl-2.0.html>.
-  ~ #L%
-  -->
-
-<EDI_Exp_Desadv AD_Client_Value="metasfresh" ReplicationEvent="0" ReplicationMode="1" ReplicationType="M" SequenceNo="1000004" TrxName="" Version="*">
-    <C_BPartner_ID>
-        <!--Bezeichnet einen Geschäftspartner-->
-        <EdiRecipientGLN>1234567890123</EdiRecipientGLN>
-        <Name>Test Kunde 1</Name>
-        <Value>G0001</Value>
-    </C_BPartner_ID>
-    <C_BPartner_Location_ID>
-        <!--Identifiziert die (Liefer-) Adresse des Geschäftspartners-->
-        <GLN>1234567890123</GLN>
-        <Name>Testadresse 3</Name>
-        <C_Location_ID>
-            <!--Location or Address-->
-            <Address1>teststraße 996</Address1>
-            <C_Country_ID>
-                <!--Country -->
-                <CountryCode>DE</CountryCode>
-            </C_Country_ID>
-            <City>teststadt 3</City>
-            <Postal>45678</Postal>
-        </C_Location_ID>
-    </C_BPartner_Location_ID>
-    <DateOrdered>2022-07-04+03:00</DateOrdered>
-    <DocumentNo>1013014</DocumentNo>
-    <EDI_Desadv_ID>1000003</EDI_Desadv_ID>
-    <EDI_ExportStatus>D</EDI_ExportStatus>
-    <MovementDate>2022-07-04+03:00</MovementDate>
-    <POReference>testReference3</POReference>
-    <EDI_Exp_Desadv_Pack>
-        <IPA_SSCC18>000000000000000062</IPA_SSCC18>
-        <M_HU_PackagingCode_LU_Text>ISO</M_HU_PackagingCode_LU_Text>
-        <GTIN_LU_PackingMaterial>EURO</GTIN_LU_PackingMaterial>
-        <EDI_Exp_Desadv_Pack_Item>
-            <EDI_Desadv_Pack_ID>1000005</EDI_Desadv_Pack_ID>
-            <EDI_DesadvLine_ID>1000003</EDI_DesadvLine_ID>
-            <QtyCU>10</QtyCU>
-            <QtyCUsPerLU>90</QtyCUsPerLU>
-            <QtyTU>9</QtyTU>
-            <EDI_DesadvLine_ID>
-                <EDI_DesadvLine_ID>1000003</EDI_DesadvLine_ID>
-                <C_UOM_ID>
-                    <!--Maßeinheit-->
-                    <Name>Stück</Name>
-                    <UOMSymbol>Stk</UOMSymbol>
-                    <X12DE355>PCE</X12DE355>
-                </C_UOM_ID>
-                <Line>20</Line>
-                <M_Product_ID>
-                    <!--Produkt, Leistung, Artikel-->
-                    <Name>Product 1</Name>
-                    <Value>P002737</Value>
-                    <Volume>0</Volume>
-                    <Weight>0.250000000000</Weight>
-                </M_Product_ID>
-                <ProductDescription>Product 1</ProductDescription>
-                <QtyEntered>100</QtyEntered>
-                <IsSubsequentDeliveryPlanned>N</IsSubsequentDeliveryPlanned>
-                <PriceActual>2.00</PriceActual>
-                <EanCom_Invoice_UOM>PCE</EanCom_Invoice_UOM>
-                <InvoicableQtyBasedOn>Nominal</InvoicableQtyBasedOn>
-                <OrderPOReference>testReference3</OrderPOReference>
-                <QtyItemCapacity>10</QtyItemCapacity>
-                <QtyDeliveredInInvoiceUOM>90</QtyDeliveredInInvoiceUOM>
-                <QtyDeliveredInUOM>90</QtyDeliveredInUOM>
-                <OrderLine>10</OrderLine>
-            </EDI_DesadvLine_ID>
-            <GTIN_TU_PackingMaterial>BPARTNER_GTIN</GTIN_TU_PackingMaterial>
-        </EDI_Exp_Desadv_Pack_Item>
-    </EDI_Exp_Desadv_Pack>
-    <Bill_Location_ID>
-        <!--Standort des Geschäftspartners für die Rechnungsstellung-->
-        <GLN>1234567890123</GLN>
-        <Name>Testadresse 3</Name>
-        <C_Location_ID>
-            <!--Location or Address-->
-            <Address1>teststraße 996</Address1>
-            <C_Country_ID>
-                <!--Country -->
-                <CountryCode>DE</CountryCode>
-            </C_Country_ID>
-            <City>teststadt 3</City>
-            <Postal>45678</Postal>
-        </C_Location_ID>
-    </Bill_Location_ID>
-    <C_Currency_ID>
-        <!--Die Währung für diesen Eintrag-->
-        <CurSymbol>€</CurSymbol>
-        <ISO_4217Numeric>978</ISO_4217Numeric>
-        <ISO_Code>EUR</ISO_Code>
-    </C_Currency_ID>
-    <HandOver_Location_ID>
-        <GLN>1234567890123</GLN>
-        <Name>Testadresse 3</Name>
-        <C_Location_ID>
-            <!--Location or Address-->
-            <Address1>teststraße 996</Address1>
-            <C_Country_ID>
-                <!--Country -->
-                <CountryCode>DE</CountryCode>
-            </C_Country_ID>
-            <City>teststadt 3</City>
-            <Postal>45678</Postal>
-        </C_Location_ID>
-    </HandOver_Location_ID>
-    <DropShip_BPartner_ID>
-        <!--Business Partner to ship to-->
-        <EdiRecipientGLN>1234567890123</EdiRecipientGLN>
-        <Name>Test Kunde 1</Name>
-        <Value>G0001</Value>
-    </DropShip_BPartner_ID>
-    <DropShip_Location_ID>
-        <!--Business Partner Location for shipping to-->
-        <GLN>1234567890123</GLN>
-        <Name>Testadresse 3</Name>
-        <C_Location_ID>
-            <!--Location or Address-->
-            <Address1>teststraße 996</Address1>
-            <C_Country_ID>
-                <!--Country -->
-                <CountryCode>DE</CountryCode>
-            </C_Country_ID>
-            <City>teststadt 3</City>
-            <Postal>45678</Postal>
-        </C_Location_ID>
-    </DropShip_Location_ID>
-    <HandOver_Partner_ID>
-        <EdiRecipientGLN>1234567890123</EdiRecipientGLN>
-        <Name>Test Kunde 1</Name>
-        <Value>G0001</Value>
-    </HandOver_Partner_ID>
-    <InvoicableQtyBasedOn>Nominal</InvoicableQtyBasedOn>
-    <DeliveryViaRule>S</DeliveryViaRule>
-    <EDI_Exp_DesadvLineWithNoPack>
-        <EDI_DesadvLine_ID>
-            <C_UOM_ID>
-                <!--Maßeinheit-->
-                <Name>Stück</Name>
-                <UOMSymbol>Stk</UOMSymbol>
-                <X12DE355>PCE</X12DE355>
-            </C_UOM_ID>
-            <EDI_DesadvLine_ID>1000000</EDI_DesadvLine_ID>
-            <Line>10</Line>
-            <M_Product_ID>
-                <!--Produkt, Leistung, Artikel-->
-                <Name>Product 2</Name>
-                <Value>P002737</Value>
-                <Volume>0</Volume>
-                <Weight>0.250000000000</Weight>
-            </M_Product_ID>
-            <ProductDescription>Product 2</ProductDescription>
-            <QtyEntered>20</QtyEntered>
-            <IsSubsequentDeliveryPlanned>N</IsSubsequentDeliveryPlanned>
-            <PriceActual>2.00</PriceActual>
-            <EanCom_Invoice_UOM>PCE</EanCom_Invoice_UOM>
-            <InvoicableQtyBasedOn>Nominal</InvoicableQtyBasedOn>
-            <OrderPOReference>poRef</OrderPOReference>
-            <QtyItemCapacity>10</QtyItemCapacity>
-            <QtyDeliveredInInvoiceUOM>0</QtyDeliveredInInvoiceUOM>
-            <QtyDeliveredInUOM>0</QtyDeliveredInUOM>
-            <OrderLine>10</OrderLine>
-        </EDI_DesadvLine_ID>
-    </EDI_Exp_DesadvLineWithNoPack>
-</EDI_Exp_Desadv>
