--- misc/services/camel/de-metas-camel-edi/src/test/resources/de/metas/edi/esb/desadvexport/stepcom/DESADV_no_packs_expected_output.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="Windows-1252" standalone="yes"?>
-<Document>
-    <Xlief4h>
-        <HEADER>
-            <DOCUMENTID>1013013</DOCUMENTID>
-            <PARTNERID>&lt;NO_DETAULT_WILL_FAIL_IF_NOT_SET&gt;</PARTNERID>
-            <OWNERID>CHEESE</OWNERID>
-            <TESTINDICATOR>T</TESTINDICATOR>
-            <APPLICATIONREF>DESADV</APPLICATIONREF>
-            <MESSAGEREF>1000004</MESSAGEREF>
-            <DOCUMENTTYP>DADV</DOCUMENTTYP>
-            <DOCUMENTFUNCTION>ORIG</DOCUMENTFUNCTION>
-            <HDATE1>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <DATEQUAL>CREA</DATEQUAL>
-                <DATEFROM>20210207213530</DATEFROM>
-            </HDATE1>
-            <HDATE1>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <DATEQUAL>DELV</DATEQUAL>
-                <DATEFROM>20220703230000</DATEFROM>
-            </HDATE1>
-            <HREFE1>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <REFERENCEQUAL>ORBU</REFERENCEQUAL>
-                <REFERENCE>testReference4</REFERENCE>
-                <REFERENCEDATE1>20220703230000</REFERENCEDATE1>
-            </HREFE1>
-            <HADRE1>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <ADDRESSQUAL>BUYR</ADDRESSQUAL>
-                <PARTYIDGLN>1234567890123</PARTYIDGLN>
-            </HADRE1>
-            <HADRE1>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <ADDRESSQUAL>DELV</ADDRESSQUAL>
-                <PARTYIDGLN>1234567890123</PARTYIDGLN>
-            </HADRE1>
-            <HADRE1>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <ADDRESSQUAL>SUPL</ADDRESSQUAL>
-                <PARTYIDGLN>9876543210987</PARTYIDGLN>
-            </HADRE1>
-            <HADRE1>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <ADDRESSQUAL>ULCO</ADDRESSQUAL>
-                <PARTYIDGLN>1234567890123</PARTYIDGLN>
-            </HADRE1>
-            <PACKIN>
-                <DOCUMENTID>1013013</DOCUMENTID>
-                <PACKAGINGTOTAL>0</PACKAGINGTOTAL>
-            </PACKIN>
-        </HEADER>
-        <TRAILR>
-            <DOCUMENTID>1013013</DOCUMENTID>
-            <CONTROLQUAL>LINE</CONTROLQUAL>
-            <CONTROLVALUE>1</CONTROLVALUE>
-        </TRAILR>
-    </Xlief4h>
-</Document>
