--- misc/services/camel/de-metas-camel-edi/src/test/resources/de/metas/edi/esb/desadvexport/stepcom/DESADV_with_and_without_packs_expected_output.xml
@@ -1,122 +1,0 @@
-<?xml version="1.0" encoding="Windows-1252" standalone="yes"?>
-<Document>
-    <Xlief4h>
-        <HEADER>
-            <DOCUMENTID>1013014</DOCUMENTID>
-            <PARTNERID>&lt;NO_DETAULT_WILL_FAIL_IF_NOT_SET&gt;</PARTNERID>
-            <OWNERID>CHEESE</OWNERID>
-            <TESTINDICATOR>T</TESTINDICATOR>
-            <APPLICATIONREF>DESADV</APPLICATIONREF>
-            <MESSAGEREF>1000004</MESSAGEREF>
-            <DOCUMENTTYP>DADV</DOCUMENTTYP>
-            <DOCUMENTFUNCTION>ORIG</DOCUMENTFUNCTION>
-            <HDATE1>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <DATEQUAL>CREA</DATEQUAL>
-                <DATEFROM>20210207213530</DATEFROM>
-            </HDATE1>
-            <HDATE1>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <DATEQUAL>DELV</DATEQUAL>
-                <DATEFROM>20220703230000</DATEFROM>
-            </HDATE1>
-            <HREFE1>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <REFERENCEQUAL>ORBU</REFERENCEQUAL>
-                <REFERENCE>testReference3</REFERENCE>
-                <REFERENCEDATE1>20220703230000</REFERENCEDATE1>
-            </HREFE1>
-            <HADRE1>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <ADDRESSQUAL>BUYR</ADDRESSQUAL>
-                <PARTYIDGLN>1234567890123</PARTYIDGLN>
-            </HADRE1>
-            <HADRE1>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <ADDRESSQUAL>DELV</ADDRESSQUAL>
-                <PARTYIDGLN>1234567890123</PARTYIDGLN>
-            </HADRE1>
-            <HADRE1>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <ADDRESSQUAL>SUPL</ADDRESSQUAL>
-                <PARTYIDGLN>9876543210987</PARTYIDGLN>
-            </HADRE1>
-            <HADRE1>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <ADDRESSQUAL>ULCO</ADDRESSQUAL>
-                <PARTYIDGLN>1234567890123</PARTYIDGLN>
-            </HADRE1>
-            <PACKIN>
-                <DOCUMENTID>1013014</DOCUMENTID>
-                <PACKAGINGTOTAL>0</PACKAGINGTOTAL>
-                <PACKAGINGLEVEL>OUTE</PACKAGINGLEVEL>
-                <PACKAGINGCODE>EURO</PACKAGINGCODE>
-                <DETAIL>
-                    <DOCUMENTID>1013014</DOCUMENTID>
-                    <LINENUMBER>10</LINENUMBER>
-                    <DPRDE1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>10</LINENUMBER>
-                        <PRODUCTDESCQUAL>PROD</PRODUCTDESCQUAL>
-                        <PRODUCTDESCTEXT>Product 2</PRODUCTDESCTEXT>
-                        <PRODUCTDESCTYPE>CU</PRODUCTDESCTYPE>
-                    </DPRDE1>
-                    <DQUAN1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>10</LINENUMBER>
-                        <QUANTITYQUAL>DELV</QUANTITYQUAL>
-                        <QUANTITY>0</QUANTITY>
-                        <MEASUREMENTUNIT>PIEC</MEASUREMENTUNIT>
-                    </DQUAN1>
-                    <DREFE1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>10</LINENUMBER>
-                        <REFERENCEQUAL>LIRN</REFERENCEQUAL>
-                        <REFERENCELINE>10</REFERENCELINE>
-                    </DREFE1>
-                    <DQVAR1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>10</LINENUMBER>
-                        <QUANTITY>-20</QUANTITY>
-                        <DISCREPANCYCODE>BCOM</DISCREPANCYCODE>
-                    </DQVAR1>
-                </DETAIL>
-                <DETAIL>
-                    <DOCUMENTID>1013014</DOCUMENTID>
-                    <LINENUMBER>20</LINENUMBER>
-                    <DPRDE1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>20</LINENUMBER>
-                        <PRODUCTDESCQUAL>PROD</PRODUCTDESCQUAL>
-                        <PRODUCTDESCTEXT>Product 1</PRODUCTDESCTEXT>
-                        <PRODUCTDESCTYPE>CU</PRODUCTDESCTYPE>
-                    </DPRDE1>
-                    <DQUAN1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>20</LINENUMBER>
-                        <QUANTITYQUAL>DELV</QUANTITYQUAL>
-                        <QUANTITY>90</QUANTITY>
-                        <MEASUREMENTUNIT>PIEC</MEASUREMENTUNIT>
-                    </DQUAN1>
-                    <DREFE1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>20</LINENUMBER>
-                        <REFERENCEQUAL>LIRN</REFERENCEQUAL>
-                        <REFERENCELINE>20</REFERENCELINE>
-                    </DREFE1>
-                    <DQVAR1>
-                        <DOCUMENTID>1013014</DOCUMENTID>
-                        <LINENUMBER>20</LINENUMBER>
-                        <QUANTITY>-10</QUANTITY>
-                        <DISCREPANCYCODE>BCOM</DISCREPANCYCODE>
-                    </DQVAR1>
-                </DETAIL>
-            </PACKIN>
-        </HEADER>
-        <TRAILR>
-            <DOCUMENTID>1013014</DOCUMENTID>
-            <CONTROLQUAL>LINE</CONTROLQUAL>
-            <CONTROLVALUE>2</CONTROLVALUE>
-        </TRAILR>
-    </Xlief4h>
-</Document>
