+++ misc/services/camel/de-metas-camel-edi/src/test/resources/de/metas/edi/esb/invoicexport/ecosio/INVOIC.xml
@@ -151,7 +151,6 @@
         <EanCom_Price_UOM>PCE</EanCom_Price_UOM>
         <EAN_CU>Qty24_CU-EAN1</EAN_CU>
         <EAN_TU>Qty24_TU-EAN1</EAN_TU>
-        <InvoicableQtyBasedOn>Nominal</InvoicableQtyBasedOn>
     </EDI_cctop_invoic_500_v>
     <EDI_cctop_invoic_500_v>
         <C_Invoice_ID>1488989</C_Invoice_ID>
@@ -178,7 +177,6 @@
         <EanCom_Price_UOM>PCE</EanCom_Price_UOM>
         <EAN_CU>Qty60_CU-EAN1</EAN_CU>
         <EAN_TU>Qty60_TU-EAN1</EAN_TU>
-        <InvoicableQtyBasedOn>Nominal</InvoicableQtyBasedOn>
     </EDI_cctop_invoic_500_v>
     <EDI_cctop_invoic_500_v>
         <C_Invoice_ID>1488989</C_Invoice_ID>
@@ -205,7 +203,6 @@
         <EanCom_Price_UOM>PCE</EanCom_Price_UOM>
         <EAN_CU>Qty10_CU-EAN1</EAN_CU>
         <EAN_TU>Qty10_TU-EAN1</EAN_TU>
-        <InvoicableQtyBasedOn>CatchWeight</InvoicableQtyBasedOn>
     </EDI_cctop_invoic_500_v>
     <EDI_cctop_901_991_v>
         <C_Invoice_ID>1488989</C_Invoice_ID>
